Created problem 14
this problem was interesting to solve and I used a dictionary containing numerical keys with a value of what the generated Collatz problem came up with previously, and when my numerical collatz journey ran into a problem indexed from that dictionary, it would stop, then append the numerical value contained + the previously found collatz value from that dictionary, and then store itself. It took me an hour-ish.